refactor: página de cadastro
