+++ src/main/webapp/css/paginas/cadastrar.css
@@ -1,4 +1,3 @@
-@import url(/css/style.css);
 @import url(/css/crud_cliente.css);
 
 #cadastrar-cliente {
