+++ src/main/webapp/paginas/cadastrar.html
@@ -1,0 +1,15 @@
+<title>Cadastro</title>
+
+<nav id="main-nav"></nav>
+
+<main id="cadastrar-cliente">
+    <div id="container-cadastrar-cliente">
+        <section id="secao-dados-pessoais"></section>
+    </div>
+
+    <button type="button" id="botao-enviar-cadastro" class="botao-salvar" onclick="enviarCliente()">
+        Salvar cliente
+    </button>
+</main>
+
+<footer id="main-footer"></footer>
