--- src/main/webapp/paginas/cliente/cadastrar/cadastrar.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Cadastro</title>
-    <link rel="stylesheet" href="/paginas/cliente/cadastrar/css/cadastrar.css">
-</head>
-
-<body>
-
-    <nav id="main-nav"></nav>
-
-    <main id="cadastrar-cliente">
-        <div id="container-cadastrar-cliente">
-            <section id="secao-dados-pessoais"></section>
-        </div>
-
-        <button type="button" id="botao-enviar-cadastro" class="botao-salvar" onclick="enviarCliente()">
-            Salvar cliente
-        </button>
-    </main>
-
-    <footer id="main-footer"></footer>
-
-    <script type="module" src="/js/loadPage.js"></script>
-    <script type="module" src="/paginas/cliente/cadastrar/js/cadastrar.js"></script>
-</body>
-
-</html>
